Add #create_team
You could break this down further into another method responsible for
creating the team.

You might be starting to see that it’s annoying that we have to keep
passing around all of the same data between each of these methods. This
is something that we’ll be able to solve when get to writing classes.